# docs/swagger.yaml
definitions:
  auth_proto.GeneratedTokens:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  handler.listUsers:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ResponseUser'
        type: array
    type: object
  model.AuthUser:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.CreateUser:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    type: object
  model.ResponseUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
  model.UpdateUser:
    properties:
      email:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
info:
  contact: {}
  description: Authenticate Service for Food Delivery Application
  title: Authenticate Service
paths:
  /user/:
    get:
      consumes:
      - application/json
      description: get list of users
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.listUsers'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: getUsers
      tags:
      - User
    post:
      consumes:
      - application/json
      description: create new user
      parameters:
      - description: User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth_proto.GeneratedTokens'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: createUser
      tags:
      - User
    put:
      consumes:
      - application/json
      description: change user password
      parameters:
      - description: User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUser'
      - description: Id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: updateUser
      tags:
      - User
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: delete user by ID
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: deleteUserByID
      tags:
      - User
    get:
      consumes:
      - application/json
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseUser'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: getUser
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: check auth information
      parameters:
      - description: User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.AuthUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth_proto.GeneratedTokens'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: authUser
      tags:
      - Auth
swagger: "2.0"